use typescript for public facing api features
