+++ platforms/hermes-javascript/.eslintrc.json
@@ -4,7 +4,8 @@
         "node": true,
         "jest/globals": true
     },
-    "plugins": ["jest"],
+    "parser": "typescript-eslint-parser",
+    "plugins": ["jest", "typescript"],
     "extends": "eslint:recommended",
     "parserOptions": {
         "ecmaVersion": 2018
@@ -12,6 +13,7 @@
     "rules": {
         "no-console": "warn",
         "quotes": ["error", "single", { "avoidEscape": true }],
-        "quote-props": ["error", "as-needed"]
+        "quote-props": ["error", "as-needed"],
+        "typescript/no-unused-vars": "off"
     }
 }
