+++ platforms/hermes-javascript/.gitignore
@@ -1,4 +1,5 @@
 .vscode
+/dist
 node_modules
 payloads
 /*.h
