+++ platforms/hermes-javascript/src/index.ts
@@ -1,5 +1,11 @@
-const Hermes = require('./api')
-const tools = require('./tools')
+// eslint-disable-next-line
+import { Hermes, HermesOptions } from './api'
+import * as tools from './tools'
+
+export * from './tools'
+export * from './api'
+
+export type Done = () => void
 
 /**
  * Sets up an event loop and initializes the Hermes class.
@@ -7,18 +13,12 @@ const tools = require('./tools')
  * @param {*} context The wrapped context function.
  * @param {*} opts Hermes options.
  */
-const withHermes = function(context, opts) {
+export const withHermes = function(context: (hermes: Hermes, done: Done) => void, opts?: HermesOptions) {
     const hermes = new Hermes(opts)
     const keepAliveRef = tools.keepAlive(20)
-    const done = () => {
+    const done: Done = () => {
         hermes.destroy()
         tools.killKeepAlive(keepAliveRef)
     }
     context(hermes, done)
-}
-
-module.exports = {
-    Hermes,
-    withHermes,
-    tools
 }
