+++ platforms/hermes-javascript/src/types/Buffer.d.ts
@@ -1,0 +1,9 @@
+export {}
+
+declare global {
+    export interface Buffer {
+        deref: () => any;
+        ref: () => Buffer;
+        isNull: () => boolean;
+    }
+}
